Fixes #6, added new Xcode 7 path for codesign_alloc
The code was refactored to loop through a list of potential paths ordered from latest to default. This should make it easier to add extra paths if it should change yet again in the future.

Much of the work was derived from [this blog post](http://blog.hoachuck.biz/blog/2013/10/29/codesign-useful-info-in-xcode-5-dot-0-1/)